+++ CHANGELOG.md
@@ -31,6 +31,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - (Batch Connect) Rescue from all Standard Exceptions and display error to user
   to keep with "app never crashes" philosophy.
 
+### Removed
+
+- Removed the ability to build a Rails app from a template as it copies from an
+  outdated template. [#215](https://github.com/OSC/ood-dashboard/issues/215)
+
 ## [1.15.2] - 2017-07-24
 
 ### Fixed